front_desk/requirements.txt
Flask==2.0.2
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-bigtable==2.4.0
google-cloud-core==2.2.1
google-cloud-datastore==2.4.0
